pm2debug_backtrace() returns an int not a size_t
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@17120 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ tbx/include/tbx_macros.h
@@ -204,7 +204,7 @@ do { \
 #  define __TBX_PRINT_TRACE()\
      ({\
        void    *array[TBX_BACKTRACE_DEPTH];\
-       size_t   size;\
+       int      size;\
 \
        size    = __TBX_RECORD_SOME_TRACE (array, TBX_BACKTRACE_DEPTH);\
        __TBX_PRINT_SOME_TRACE (array, size); \
